+++ anyblok_pyramid/__init__.py
@@ -22,33 +22,3 @@ def get_callable(k):
 
 
 from . import handler  # noqa
-
-
-def load_init_function_from_entry_points():
-    """Call all the entry point ``anyblok_pyramid.init`` to update
-    the argument setting
-
-    the callable need to have one parametter, it is a dict::
-
-        def init_function():
-            ...
-
-    We add the entry point by the setup file::
-
-        setup(
-            ...,
-            entry_points={
-                'anyblok_pyramid.init': [
-                    init_function=path:init_function,
-                    ...
-                ],
-            },
-            ...,
-        )
-
-
-    """
-    from pkg_resources import iter_entry_points
-    for i in iter_entry_points('anyblok_pyramid.init'):
-        logger.debug('Load init: %r' % i.name)
-        i.load()()